test: move all tests into tests/ directory
